display: reset DSR_COUNT to the minimum
DSR count had been increased from 2 to 15 three years ago in order to make
it less power-agressive and avoid shutting down the display for nothing.
On our platforms, we need to reduce it in order to increase the s0ix residency
in use cases with a few fps.

A previous patch already decrease this DSR count from 15 to 4.
Now let's revert completely the change made three years ago by setting it to 2

It allows to sleep as much as possible between 2 frames and avoid any
side-effect on dozing needs by requesting unexpected VSYNC reSync.

Change-Id: I595b836698e2f769cfaaabb496d6f440dbef1b0d
Tracked-On: https://jira01.devtools.intel.com/browse/AW-4084
Signed-off-by: Alexandre Ledentec <alexandrex.ledentec@intel.com>
Reviewed-on: https://android.intel.com/571897
Reviewed-by: Dubray, SimonX <simonx.dubray@intel.com>
Tested-by: Dubray, SimonX <simonx.dubray@intel.com>
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>